requirements.txt
# ──────────────────────────────────────────
# Data & ETL
# ──────────────────────────────────────────
pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0
pyodbc>=5.0

# ──────────────────────────────────────────
# Machine Learning
# ──────────────────────────────────────────
scikit-learn>=1.3
xgboost>=2.0
scipy>=1.11
python-dotenv>=1.0.0

# ──────────────────────────────────────────
# Market Basket & Recommendations
# ──────────────────────────────────────────
mlxtend>=0.23

# ──────────────────────────────────────────
# Visualization
# ──────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15

# ──────────────────────────────────────────
# Notebooks & Dashboard
# ──────────────────────────────────────────
jupyter>=1.0
notebook>=7.0

# ──────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────
python-docx>=1.0
tqdm>=4.65
